implement IP tracker cli
This cli application takes a user name as a command line argument, and then queries for users current IP and stores it in persistent storage (along other users IPs). Each time the application is ran, given users IP is persisted. Note that users name is case insensitive so if ran with "Adrian", and then with "adrian" use user name, then initial entry will be updated, instead of creating a new one.

See README.md for instructions on how to build and run.